requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==1.10.13
motor==3.3.1
pymongo==4.5.0
python-dotenv==1.0.0
google-generativeai==0.3.1
redis==5.0.1
requests==2.31.0
wikipedia-api==0.5.8
aiohttp==3.9.1
python-multipart==0.0.6
pytest==7.4.3
httpx==0.25.1 
